Update idl_gen test
