--- tests/idl_gen/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    mantle::build_service().unwrap();
-}
