+++ tests/idl_gen/res/TestService.json
@@ -1,6 +1,6 @@
 {
   "name": "TestService",
-  "namespace": "test_mantle_build",
+  "namespace": "test_idl_gen",
   "imports": [
     {
       "name": "testlib",
@@ -27,8 +27,7 @@
             "params": {
               "type": "DefTy"
             }
-          },
-          "indexed": true
+          }
         },
         {
           "name": "non_indexed",
@@ -97,10 +96,10 @@
             "type": "tuple",
             "params": [
               {
-                "type": "h256"
+                "type": "address"
               },
               {
-                "type": "u256"
+                "type": "u64"
               },
               {
                 "type": "address"
@@ -118,15 +117,13 @@
           "name": "indexed1",
           "type": {
             "type": "u32"
-          },
-          "indexed": true
+          }
         },
         {
           "name": "indexed2",
           "type": {
             "type": "u32"
-          },
-          "indexed": true
+          }
         }
       ]
     }
